refactor: extract data reducer into a testable module + add tests
The Redux `data` reducer (the core notes/folders/storages/tags state
machine) lived inside store.js, which transitively imports electron via
ConfigManager and so cannot be imported in jest -- leaving the most
important state logic completely untested.

Move the `data` reducer, its pure helpers (updateStarredChange,
updateFolderChange, updateTagChanges, assignToTags, removeFromTags,
getOrInitItem) and defaultDataMap verbatim into browser/main/dataReducer.js
and import it back into store.js. No behavior change -- it is a pure move;
store.js keeps the config/status reducers and the store wiring.

Add tests for the reducer: defaultDataMap, unknown-action passthrough,
INIT_ALL, ADD_STORAGE, UPDATE_NOTE (insert + index), trashing a note,
DELETE_NOTE, and previous-state immutability.

Verified: webpack production build still compiles (the store import wiring
is intact), full suite green (193 -> 201 passing), lint clean.
